# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents/pdf:
    post:
      operationId: renderSamplePdf
      tags: [documents]
      summary: Render a safe sample document PDF
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SampleDocumentInput"
      responses:
        "202":
          description: PDF export boundary accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfExportResponse"
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Start an admin session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSessionResponse"
        "401":
          description: Invalid admin credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/admin/session:
    get:
      operationId: getAdminSession
      tags: [auth]
      summary: Get the current admin session
      responses:
        "200":
          description: Current admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSessionResponse"
        "401":
          description: Admin session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/token/login:
    post:
      operationId: tokenLogin
      tags: [auth]
      summary: Sign in with an issued customer access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenLoginInput"
      responses:
        "200":
          description: Customer session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSessionResponse"
        "401":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: Token temporarily locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/session:
    get:
      operationId: getUserSession
      tags: [auth]
      summary: Get the current customer session
      responses:
        "200":
          description: Current customer session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSessionResponse"
        "401":
          description: Customer session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUserAccount
      tags: [auth]
      summary: Soft-delete the current customer account
      responses:
        "204":
          description: Customer account deactivated
        "401":
          description: Customer session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payees:
    get:
      operationId: listPayees
      tags: [payees]
      summary: List saved payee profiles
      responses:
        "200":
          description: Saved payees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayeeListResponse"
        "401":
          description: Customer session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createPayee
      tags: [payees]
      summary: Save a payee profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayeeInput"
      responses:
        "201":
          description: Payee saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payee"
        "401":
          description: Customer session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payees/{payeeId}:
    delete:
      operationId: deletePayee
      tags: [payees]
      summary: Delete a saved payee profile
      parameters:
        - $ref: "#/components/parameters/PayeeId"
      responses:
        "204":
          description: Payee deleted
        "401":
          description: Customer session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Payee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders:
    post:
      operationId: createOrder
      tags: [commerce]
      summary: Create a manual-payment order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/lookup:
    post:
      operationId: lookupOrder
      tags: [commerce]
      summary: Track an order by order number and email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderLookupInput"
      responses:
        "200":
          description: Order status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/usage:
    post:
      operationId: recordDocumentUsage
      tags: [documents]
      summary: Record a PDF or browser print usage event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageInput"
      responses:
        "200":
          description: Updated usage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageResponse"
        "401":
          description: Customer session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Usage limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    PayeeId:
      name: payeeId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    AdminLoginInput:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 120
        password:
          type: string
          minLength: 1
          maxLength: 256
      required: [username, password]
    AdminSessionResponse:
      type: object
      properties:
        authenticated:
          type: boolean
        role:
          type: string
          enum: [admin]
      required: [authenticated, role]
    TokenLoginInput:
      type: object
      properties:
        token:
          type: string
          minLength: 16
          maxLength: 128
      required: [token]
    UserSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [active, restricted, deleted]
        slipLimit:
          type: number
        slipsUsed:
          type: number
        designLimit:
          type: number
        designsUsed:
          type: number
        pdfPrintCount:
          type: number
        browserPrintCount:
          type: number
        lockoutUntil:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, email, status, slipLimit, slipsUsed, designLimit, designsUsed, pdfPrintCount, browserPrintCount, lockoutUntil, createdAt]
...
    UserSessionResponse:
      type: object
      properties:
        authenticated:
          type: boolean
        user:
          $ref: "#/components/schemas/UserSummary"
      required: [authenticated, user]
    PayeeInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        address:
          type: string
          maxLength: 200
        bankName:
          type: string
          maxLength: 100
        bankAddress:
          type: string
          maxLength: 200
        routingNumber:
          type: string
          pattern: "^[0-9]{9}$"
        accountNumber:
          type: string
          maxLength: 30
        bankLogoUrl:
          type: ["string", "null"]
          maxLength: 500
      required: [name, address, bankName, bankAddress, routingNumber, accountNumber, bankLogoUrl]
    Payee:
      allOf:
        - $ref: "#/components/schemas/PayeeInput"
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
          required: [id, createdAt]
    PayeeListResponse:
      type: array
      items:
        $ref: "#/components/schemas/Payee"
...
    CreateOrderInput:
      type: object
      properties:
        planId:
          type: string
          enum: [starter, pro, enterprise]
        name:
          type: string
          minLength: 1
          maxLength: 120
        email:
          type: string
        paymentMethod:
          type: string
          enum: [card_manual, bitcoin]
        acceptedPaymentTerms:
          type: boolean
      required: [planId, name, email, paymentMethod, acceptedPaymentTerms]
    OrderLookupInput:
      type: object
      properties:
        orderNumber:
          type: string
          minLength: 6
          maxLength: 40
        email:
          type: string
      required: [orderNumber, email]
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        planId:
          type: string
          enum: [starter, pro, enterprise]
        name:
          type: string
        email:
          type: string
        paymentMethod:
          type: string
          enum: [card_manual, bitcoin]
        status:
          type: string
          enum: [pending, successful, failed]
        amountCents:
          type: number
        currency:
          type: string
        acceptedPaymentTerms:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, orderNumber, planId, name, email, paymentMethod, status, amountCents, currency, acceptedPaymentTerms, createdAt, updatedAt]
    UsageInput:
      type: object
      properties:
        kind:
          type: string
          enum: [pdf, browser_print]
      required: [kind]
    UsageResponse:
      type: object
      properties:
        kind:
          type: string
          enum: [pdf, browser_print]
        slipsUsed:
          type: number
        slipLimit:
          type: number
        remaining:
          type: number
      required: [kind, slipsUsed, slipLimit, remaining]
...
    SampleDocumentInput:
      type: object
      properties:
        payeeName:
          type: string
          maxLength: 100
        payeeAddress:
          type: string
          maxLength: 200
        payorName:
          type: string
          maxLength: 100
        payorAddress:
          type: string
          maxLength: 200
        bankName:
          type: string
          maxLength: 100
        bankAddress:
          type: string
          maxLength: 200
        bankLogoDataUrl:
          type: ["string", "null"]
          maxLength: 4194304
        accountNumber:
          type: string
          maxLength: 30
        routingNumber:
          type: string
          pattern: "^[0-9]{9}$"
        checkNumber:
          type: string
          maxLength: 20
        date:
          type: string
          format: date
        amount:
          type: ["number", "null"]
        memo:
          type: string
          maxLength: 200
      required: [payeeName, bankAddress, accountNumber, routingNumber, checkNumber, date, amount, memo]
    PdfExportResponse:
      type: object
      properties:
        status:
          type: string
          enum: [not_configured, ready]
        message:
          type: string
      required: [status, message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]