# lib/api-zod/src/generated/types/orderLookupInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderLookupInput {
  /**
     * @minLength 6
     * @maxLength 40
     */
  orderNumber: string;
  email: string;
}