# lib/api-zod/src/generated/types/payeeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PayeeInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /** @maxLength 200 */
  address: string;
  /** @maxLength 100 */
  bankName: string;
  /** @maxLength 200 */
  bankAddress: string;
  /** @pattern ^[0-9]{9}$ */
  routingNumber: string;
  /** @maxLength 30 */
  accountNumber: string;
  /**
     * @maxLength 500
     * @nullable
     */
  bankLogoUrl: string | null;
}