# lib/api-zod/src/generated/types/usageResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UsageResponseKind } from './usageResponseKind';

export interface UsageResponse {
  kind: UsageResponseKind;
  slipsUsed: number;
  slipLimit: number;
  remaining: number;
}